#268: This changes the PP Controller to only return studies in an active or paused state. Also refactored the Controller to a single Controller implementation that implements both APIs to better share common functionality.
